error handling

+++ scripts/cli/src/fully_managed_connector_show_config_command.sh
@@ -62,31 +62,45 @@ do
         fi
         curl_output=$(curl $security -s -X GET -H "Content-Type: application/json" "https://api.confluent.cloud/connect/v1/environments/$environment/clusters/$cluster/connectors/$connector/config" --header "authorization: Basic $authorization")
         ret=$?
-        set -e
         if [ $ret -eq 0 ]
         then
-            echo "playground fully-managed-connector create-or-update --connector $connector << EOF"
-            echo "$curl_output" | jq -S . | sed 's/\$/\\$/g'
-            echo "EOF"
-
-            if [[ "$OSTYPE" == "darwin"* ]]
+            if echo "$curl_output" | jq 'if .error then .error | has("code") else has("error_code") end' 2> /dev/null | grep -q true
             then
-                clipboard=$(playground config get clipboard)
-                if [ "$clipboard" == "" ]
+                if echo "$curl_output" | jq '.error | has("code")' 2> /dev/null | grep -q true
                 then
-                    playground config set clipboard true
+                    code=$(echo "$curl_output" | jq -r .error.code)
+                    message=$(echo "$curl_output" | jq -r .error.message)
+                else
+                    code=$(echo "$curl_output" | jq -r .error_code)
+                    message=$(echo "$curl_output" | jq -r .message)
                 fi
+                logerror "Command failed with error code $code"
+                logerror "$message"
+                exit 1
+            else
+                echo "playground fully-managed-connector create-or-update --connector $connector << EOF"
+                echo "$curl_output" | jq -S . | sed 's/\$/\\$/g'
+                echo "EOF"
 
-                if [ "$clipboard" == "true" ] || [ "$clipboard" == "" ]
+                if [[ "$OSTYPE" == "darwin"* ]]
                 then
-                    tmp_dir=$(mktemp -d -t ci-XXXXXXXXXX)
-                    trap 'rm -rf $tmp_dir' EXIT
-                    echo "playground fully-managed-connector --connector $connector << EOF" > $tmp_dir/tmp
-                    echo "$curl_output" | jq -S . | sed 's/\$/\\$/g' >> $tmp_dir/tmp
-                    echo "EOF" >> $tmp_dir/tmp
+                    clipboard=$(playground config get clipboard)
+                    if [ "$clipboard" == "" ]
+                    then
+                        playground config set clipboard true
+                    fi
 
-                    cat $tmp_dir/tmp | pbcopy
-                    log "📋 connector config has been copied to the clipboard (disable with 'playground config set clipboard false')"
+                    if [ "$clipboard" == "true" ] || [ "$clipboard" == "" ]
+                    then
+                        tmp_dir=$(mktemp -d -t ci-XXXXXXXXXX)
+                        trap 'rm -rf $tmp_dir' EXIT
+                        echo "playground fully-managed-connector --connector $connector << EOF" > $tmp_dir/tmp
+                        echo "$curl_output" | jq -S . | sed 's/\$/\\$/g' >> $tmp_dir/tmp
+                        echo "EOF" >> $tmp_dir/tmp
+
+                        cat $tmp_dir/tmp | pbcopy
+                        log "📋 connector config has been copied to the clipboard (disable with 'playground config set clipboard false')"
+                    fi
                 fi
             fi
         else
